requirements.txt
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.5.0
pydocstyle==5.0.2
# pyflakes==2.2.0
# pylama==7.7.1
pytz==2020.1
snowballstemmer==2.0.0
sqlparse==0.3.1
